More granular loss testing.
